Test Backward Euler by MG
Tidy up of the code with structure. Added the arguments T_comput and showAndSave to the function simultaion, allowing to prevent the program to plot the fields and computing longer than a given time. Finally tests were realised and outputs pictures are in the folder ./outputs_program_backwardEuler . They show the algorithm is out of service for now and needs correction.